# add test inspector as separate project and add command line option to enable it on demand