--- a/test/common/make-testinsp
+++ b/test/common/make-testinsp
@@ -1,0 +1,17 @@
+#!/bin/sh
+
+# prepare testinspector
+GITHUB_REPO='testinsp'
+SUBDIR='testinsp'
+COMMON=$(dirname $0)
+V="${V-0}" # default to friendly messages
+
+set -eu
+cd $COMMON
+. ./git-utils.sh
+
+if [ ! -e "$COMMON/$SUBDIR" ]; then
+    git clone https://github.com/jscotka/testinsp.git $SUBDIR
+else
+    echo "$SUBDIR/ already exists, skipping"
+fi