--- a/test/common/make-testinsp
+++ b/test/common/make-testinsp
@@ -1,7 +1,6 @@
 #!/bin/sh
 
 # prepare testinspector
-GITHUB_REPO='testinsp'
 SUBDIR='testinsp'
 COMMON=$(dirname $0)
 V="${V-0}" # default to friendly messages