--- a/test/common/make-testinsp
+++ b/test/common/make-testinsp
@@ -1,16 +1,16 @@
 #!/bin/sh
 
 # prepare testinspector
+
 SUBDIR='testinsp'
 COMMON=$(dirname $0)
 V="${V-0}" # default to friendly messages
 
 set -eu
 cd $COMMON
-. ./git-utils.sh
 
 if [ ! -e "$COMMON/$SUBDIR" ]; then
-    git clone https://github.com/jscotka/testinsp.git $SUBDIR
+    git clone https://github.com/jscotka/${SUBDIR}.git $SUBDIR
 else
     echo "$SUBDIR/ already exists, skipping"
 fi