@jscotka 2024-01-05
another variant of https://github.com/cockpit-project/cockpit/pull/19756