@martinpitt 2024-12-16
See https://github.com/cockpit-project/cockpit/pull/19756#issuecomment-2544870178